adapted to PR review
