+++ src/builder/pod/volume.rs
@@ -1,3 +1,4 @@
+use k8s_openapi::api::core::v1::VolumeMount;
 use k8s_openapi::{
     api::core::v1::{
         CSIVolumeSource, ConfigMapVolumeSource, DownwardAPIVolumeSource, EmptyDirVolumeSource,
@@ -6,6 +7,7 @@ use k8s_openapi::{
     },
     apimachinery::pkg::api::resource::Quantity,
 };
+use std::collections::BTreeMap;
 
 /// A builder to build [`Volume`] objects.
 /// May only contain one `volume_source` at a time.
@@ -189,9 +191,132 @@ impl VolumeBuilder {
     }
 }
 
+/// A builder to build [`VolumeMount`] objects.
+///
+#[derive(Clone, Default)]
+pub struct VolumeMountBuilder {
+    mount_path: String,
+    mount_propagation: Option<String>,
+    name: String,
+    read_only: Option<bool>,
+    sub_path: Option<String>,
+    sub_path_expr: Option<String>,
+}
+
+impl VolumeMountBuilder {
+    pub fn new(name: impl Into<String>, mount_path: impl Into<String>) -> VolumeMountBuilder {
+        VolumeMountBuilder {
+            mount_path: mount_path.into(),
+            name: name.into(),
+            ..VolumeMountBuilder::default()
+        }
+    }
+
+    pub fn read_only(&mut self, read_only: bool) -> &mut Self {
+        self.read_only = Some(read_only);
+        self
+    }
+
+    pub fn mount_propagation(&mut self, mount_propagation: impl Into<String>) -> &mut Self {
+        self.mount_propagation = Some(mount_propagation.into());
+        self
+    }
+
+    pub fn sub_path(&mut self, sub_path: impl Into<String>) -> &mut Self {
+        self.sub_path = Some(sub_path.into());
+        self
+    }
+
+    pub fn sub_path_expr(&mut self, sub_path_expr: impl Into<String>) -> &mut Self {
+        self.sub_path_expr = Some(sub_path_expr.into());
+        self
+    }
+
+    /// Consumes the Builder and returns a constructed VolumeMount
+    pub fn build(&self) -> VolumeMount {
+        VolumeMount {
+            mount_path: self.mount_path.clone(),
+            mount_propagation: self.mount_propagation.clone(),
+            name: self.name.clone(),
+            read_only: self.read_only,
+            sub_path: self.sub_path.clone(),
+            sub_path_expr: self.sub_path_expr.clone(),
+        }
+    }
+}
+
+#[derive(Clone)]
+pub struct SecretOperatorVolumeSourceBuilder {
+    secret_class: String,
+    scopes: Vec<SecretOperatorVolumeScope>,
+}
+
+impl SecretOperatorVolumeSourceBuilder {
+    pub fn new(secret_class: impl Into<String>) -> Self {
+        Self {
+            secret_class: secret_class.into(),
+            scopes: Vec::new(),
+        }
+    }
+
+    pub fn with_node_scope(&mut self) -> &mut Self {
+        self.scopes.push(SecretOperatorVolumeScope::Node);
+        self
+    }
+
+    pub fn with_pod_scope(&mut self) -> &mut Self {
+        self.scopes.push(SecretOperatorVolumeScope::Pod);
+        self
+    }
+
+    pub fn with_service_scope(&mut self, name: impl Into<String>) -> &mut Self {
+        self.scopes
+            .push(SecretOperatorVolumeScope::Service { name: name.into() });
+        self
+    }
+
+    pub fn build(&self) -> CSIVolumeSource {
+        let mut attrs = BTreeMap::from([(
+            "secrets.stackable.tech/class".to_string(),
+            self.secret_class.clone(),
+        )]);
+
+        if !self.scopes.is_empty() {
+            let mut scopes = String::new();
+            for scope in self.scopes.iter() {
+                if !scopes.is_empty() {
+                    scopes.push(',');
+                };
+                match scope {
+                    SecretOperatorVolumeScope::Node => scopes.push_str("node"),
+                    SecretOperatorVolumeScope::Pod => scopes.push_str("pod"),
+                    SecretOperatorVolumeScope::Service { name } => {
+                        scopes.push_str("service=");
+                        scopes.push_str(name);
+                    }
+                }
+            }
+            attrs.insert("secrets.stackable.tech/scope".to_string(), scopes);
+        }
+
+        CSIVolumeSource {
+            driver: "secrets.stackable.tech".to_string(),
+            volume_attributes: Some(attrs),
+            ..CSIVolumeSource::default()
+        }
+    }
+}
+
+#[derive(Clone)]
+enum SecretOperatorVolumeScope {
+    Node,
+    Pod,
+    Service { name: String },
+}
+
 #[cfg(test)]
 mod tests {
-    use crate::builder::pod::volume::VolumeBuilder;
+    use super::*;
     use k8s_openapi::apimachinery::pkg::api::resource::Quantity;
 
     #[test]
@@ -229,5 +354,24 @@ mod tests {
             vol.secret.and_then(|secret| secret.secret_name),
             Some("secret".to_string())
         );
+    }
+
+    #[test]
+    fn test_volume_mount_builder() {
+        let mut volume_mount_builder = VolumeMountBuilder::new("name", "mount_path");
+        volume_mount_builder
+            .mount_propagation("mount_propagation")
+            .read_only(true)
+            .sub_path("sub_path")
+            .sub_path_expr("sub_path_expr");
+
+        let vm = volume_mount_builder.build();
+
+        assert_eq!(vm.name, "name".to_string());
+        assert_eq!(vm.mount_path, "mount_path".to_string());
+        assert_eq!(vm.mount_propagation, Some("mount_propagation".to_string()));
+        assert_eq!(vm.read_only, Some(true));
+        assert_eq!(vm.sub_path, Some("sub_path".to_string()));
+        assert_eq!(vm.sub_path_expr, Some("sub_path_expr".to_string()));
     }
 }
